flask_runpod_app/requirements.txt
# Flask and Web Framework
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7

# Database
SQLAlchemy==2.0.21

# Server
gunicorn==21.2.0

# Configuration
python-dotenv==1.0.0

# Cloud Storage (Cloudflare R2)
boto3==1.28.57
botocore==1.31.57

# RunPod API (optional, for API-based management)
# runpod==1.3.0  # Uncomment if using RunPod Python SDK

# Email
email-validator==2.0.0

# Data Processing
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2

# Video Processing (for frame counting)
opencv-python-headless==4.8.1.78

# HTTP Requests
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3